Accpet client-side send and recieve timeouts. [#CLICKHOUSE-2]
